+++ drivers/mysql/mysql.go
@@ -8,6 +8,8 @@ import (
 	"slices"
 	"strings"
 
+	"github.com/neilotoole/sq/libsq/ast"
+
 	"github.com/neilotoole/sq/libsq/core/tablefq"
 
 	"github.com/neilotoole/sq/libsq/core/loz"
@@ -137,7 +139,9 @@ func placeholders(numCols, numRows int) string {
 
 // Renderer implements driver.SQLDriver.
 func (d *driveri) Renderer() *render.Renderer {
-	return render.NewDefaultRenderer()
+	r := render.NewDefaultRenderer()
+	r.FunctionNames[ast.FuncNameSchema] = "DATABASE"
+	return r
 }
 
 // RecordMeta implements driver.SQLDriver.