+++ drivers/sqlite3/sqlite3.go
@@ -16,6 +16,8 @@ import (
 	"sync"
 	"time"
 
+	"github.com/neilotoole/sq/libsq/ast"
+
 	"github.com/neilotoole/sq/libsq/core/tablefq"
 
 	"github.com/neilotoole/sq/libsq/core/loz"
@@ -239,7 +241,9 @@ func placeholders(numCols, numRows int) string {
 
 // Renderer implements driver.SQLDriver.
 func (d *driveri) Renderer() *render.Renderer {
-	return render.NewDefaultRenderer()
+	r := render.NewDefaultRenderer()
+	r.FunctionOverrides[ast.FuncNameSchema] = doRenderFuncSchema
+	return r
 }
 
 // CopyTable implements driver.SQLDriver.