+++ libsq/ast/func.go
@@ -6,6 +6,16 @@ import (
 	"github.com/neilotoole/sq/libsq/ast/internal/slq"
 )
 
+const (
+	FuncNameAvg         = "avg"
+	FuncNameCount       = "count"
+	FuncNameCountUnique = "count_unique"
+	FuncNameMax         = "max"
+	FuncNameMin         = "min"
+	FuncNameSchema      = "schema"
+	FuncNameSum         = "sum"
+)
+
 var (
 	_ Node         = (*FuncNode)(nil)
 	_ ResultColumn = (*FuncNode)(nil)